+++ pyconkr/context_processors.py
@@ -6,7 +6,9 @@ from django.contrib.flatpages.models import FlatPage
 from django.db.models import Count
 from django.utils.translation import ugettext_lazy as _
 
-from .models import SponsorLevel, Speaker, Banner
+from pyconkr.models import Banner
+from sponsor.models import SponsorLevel
+from program.models import Speaker
 
 
 def default(request):
